+++ lib/robots_txt.dart
@@ -1,5 +1,6 @@
 /// Lightweight, fully documented `robots.txt` file parser.
 library robots_txt;
 
-export 'src/parser.dart';
-export 'src/rule.dart';
+export 'src/parser.dart' show Robots, PrecedentRuleType, FieldType;
+export 'src/rule.dart' show Rule, FindRule, Precedence, PrecedenceStrategy;
+export 'src/ruleset.dart' show Ruleset, FindRuleInRuleset;